Refactor for log streaming query params
